1) Port to NUC970 2) Some warnings fixed

+++ src/m_misc.c
@@ -1,4 +1,4 @@
-// Emacs style mode select   -*- C++ -*- 
+// Emacs style mode select   -*- C++ -*-
 //-----------------------------------------------------------------------------
 //
 // $Id:$
@@ -85,7 +85,7 @@ M_DrawText
 	    x += 4;
 	    continue;
 	}
-		
+
 	w = SHORT (hu_font[c]->width);
 	if (x+w > SCREENWIDTH)
 	    break;
@@ -117,7 +117,7 @@ M_WriteFile
 {
     int		handle;
     int		count;
-	
+
     handle = open ( name, O_WRONLY | O_CREAT | O_TRUNC | O_BINARY, 0666);
 
     if (handle == -1)
@@ -125,10 +125,10 @@ M_WriteFile
 
     count = write (handle, source, length);
     close (handle);
-	
+
     if (count < length)
 	return false;
-		
+
     return true;
 }
 
@@ -144,7 +144,7 @@ M_ReadFile
     int	handle, count, length;
     struct stat	fileinfo;
     byte		*buf;
-	
+
     handle = open (name, O_RDONLY | O_BINARY, 0666);
     if (handle == -1)
 	I_Error ("Couldn't read file %s", name);
@@ -154,10 +154,10 @@ M_ReadFile
     buf = Z_Malloc (length, PU_STATIC, NULL);
     count = read (handle, buf, length);
     close (handle);
-	
+
     if (count < length)
 	I_Error ("Couldn't read file %s", name);
-		
+
     *buffer = buf;
     return length;
 }
@@ -226,7 +226,7 @@ typedef struct
 {
     char*	name;
     int*	location;
-    int64_t	defaultvalue;
+    intptr_t	defaultvalue;
     int		scantranslate;		// PC scan code hack
     int		untranslated;		// lousy hack
 } default_t;
@@ -237,7 +237,7 @@ default_t	defaults[] =
     {"sfx_volume",&snd_SfxVolume, 8},
     {"music_volume",&snd_MusicVolume, 8},
     {"show_messages",&showMessages, 1},
-    
+
 
 #ifdef NORMALUNIX
     {"key_right",&key_right, KEY_RIGHTARROW},
@@ -252,17 +252,17 @@ default_t	defaults[] =
     {"key_strafe",&key_strafe, KEY_RALT},
     {"key_speed",&key_speed, KEY_RSHIFT},
 
-// UNIX hack, to be removed. 
+// UNIX hack, to be removed.
 #ifdef SNDSERV
-    {"sndserver", (int *) &sndserver_filename, (int64_t) "sndserver"},
+    {"sndserver", (int *) &sndserver_filename, (intptr_t) "sndserver"},
     {"mb_used", &mb_used, 2},
 #endif
-    
+
 #endif
 
 #ifdef LINUX
-    {"mousedev", (int*)&mousedev, (int)"/dev/ttyS0"},
-    {"mousetype", (int*)&mousetype, (int)"microsoft"},
+    {"mousedev", (int*)&mousedev, (intptr_t)"/dev/ttyS0"},
+    {"mousetype", (int*)&mousetype, (intptr_t)"microsoft"},
 #endif
 
     {"use_mouse",&usemouse, 1},
@@ -285,16 +285,16 @@ default_t	defaults[] =
 
     {"usegamma",&usegamma, 0},
 
-    {"chatmacro0", (int *) &chat_macros[0], (int64_t) HUSTR_CHATMACRO0 },
-    {"chatmacro1", (int *) &chat_macros[1], (int64_t) HUSTR_CHATMACRO1 },
-    {"chatmacro2", (int *) &chat_macros[2], (int64_t) HUSTR_CHATMACRO2 },
-    {"chatmacro3", (int *) &chat_macros[3], (int64_t) HUSTR_CHATMACRO3 },
-    {"chatmacro4", (int *) &chat_macros[4], (int64_t) HUSTR_CHATMACRO4 },
-    {"chatmacro5", (int *) &chat_macros[5], (int64_t) HUSTR_CHATMACRO5 },
-    {"chatmacro6", (int *) &chat_macros[6], (int64_t) HUSTR_CHATMACRO6 },
-    {"chatmacro7", (int *) &chat_macros[7], (int64_t) HUSTR_CHATMACRO7 },
-    {"chatmacro8", (int *) &chat_macros[8], (int64_t) HUSTR_CHATMACRO8 },
-    {"chatmacro9", (int *) &chat_macros[9], (int64_t) HUSTR_CHATMACRO9 }
+    {"chatmacro0", (int *) &chat_macros[0], (intptr_t) HUSTR_CHATMACRO0 },
+    {"chatmacro1", (int *) &chat_macros[1], (intptr_t) HUSTR_CHATMACRO1 },
+    {"chatmacro2", (int *) &chat_macros[2], (intptr_t) HUSTR_CHATMACRO2 },
+    {"chatmacro3", (int *) &chat_macros[3], (intptr_t) HUSTR_CHATMACRO3 },
+    {"chatmacro4", (int *) &chat_macros[4], (intptr_t) HUSTR_CHATMACRO4 },
+    {"chatmacro5", (int *) &chat_macros[5], (intptr_t) HUSTR_CHATMACRO5 },
+    {"chatmacro6", (int *) &chat_macros[6], (intptr_t) HUSTR_CHATMACRO6 },
+    {"chatmacro7", (int *) &chat_macros[7], (intptr_t) HUSTR_CHATMACRO7 },
+    {"chatmacro8", (int *) &chat_macros[8], (intptr_t) HUSTR_CHATMACRO8 },
+    {"chatmacro9", (int *) &chat_macros[9], (intptr_t) HUSTR_CHATMACRO9 }
 
 };
 
@@ -310,11 +310,11 @@ void M_SaveDefaults (void)
     int		i;
     int		v;
     FILE*	f;
-	
+
     f = fopen (defaultfile, "w");
     if (!f)
 	return; // can't write the file, but don't complain
-		
+
     for (i=0 ; i<numdefaults ; i++)
     {
 	if (defaults[i].defaultvalue > -0xfff
@@ -327,7 +327,7 @@ void M_SaveDefaults (void)
 		     * (char **) (defaults[i].location));
 	}
     }
-	
+
     fclose (f);
 }
 
@@ -347,12 +347,12 @@ void M_LoadDefaults (void)
     char*	newstring;
     int		parm;
     boolean	isstring;
-    
+
     // set everything to base values
     numdefaults = sizeof(defaults)/sizeof(defaults[0]);
     for (i=0 ; i<numdefaults ; i++)
 	*defaults[i].location = defaults[i].defaultvalue;
-    
+
     // check for a custom default file
     i = M_CheckParm ("-config");
     if (i && i<myargc-1)
@@ -362,7 +362,7 @@ void M_LoadDefaults (void)
     }
     else
 	defaultfile = basedefault;
-    
+
     // read the file in, overriding any set defaults
     f = fopen (defaultfile, "r");
     if (f)
@@ -397,7 +397,7 @@ void M_LoadDefaults (void)
 		    }
 	    }
 	}
-		
+
 	fclose (f);
     }
 }
@@ -419,17 +419,17 @@ typedef struct
     unsigned short	ymin;
     unsigned short	xmax;
     unsigned short	ymax;
-    
+
     unsigned short	hres;
     unsigned short	vres;
 
     unsigned char	palette[48];
-    
+
     char		reserved;
     char		color_planes;
     unsigned short	bytes_per_line;
     unsigned short	palette_type;
-    
+
     char		filler[58];
     unsigned char	data;		// unbounded
 } pcx_t;
@@ -450,7 +450,7 @@ WritePCXfile
     int		length;
     pcx_t*	pcx;
     byte*	pack;
-	
+
     pcx = Z_Malloc (width*height*2+1000, PU_STATIC, NULL);
 
     pcx->manufacturer = 0x0a;		// PCX id
@@ -472,7 +472,7 @@ WritePCXfile
 
     // pack the image
     pack = &pcx->data;
-	
+
     for (i=0 ; i<width*height ; i++)
     {
 	if ( (*data & 0xc0) != 0xc0)
@@ -483,12 +483,12 @@ WritePCXfile
 	    *pack++ = *data++;
 	}
     }
-    
+
     // write the palette
     *pack++ = 0x0c;	// palette ID byte
     for (i=0 ; i<768 ; i++)
 	*pack++ = *palette++;
-    
+
     // write output file
     length = pack - (byte *)pcx;
     M_WriteFile (filename, pcx, length);
@@ -505,14 +505,14 @@ void M_ScreenShot (void)
     int		i;
     byte*	linear;
     char	lbmname[12];
-    
+
     // munge planar buffer to linear
     linear = screens[2];
     I_ReadScreen (linear);
-    
+
     // find a file name to save it to
     strcpy(lbmname,"DOOM00.pcx");
-		
+
     for (i=0 ; i<=99 ; i++)
     {
 	lbmname[4] = i/10 + '0';
@@ -522,12 +522,12 @@ void M_ScreenShot (void)
     }
     if (i==100)
 	I_Error ("M_ScreenShot: Couldn't create a PCX");
-    
+
     // save the pcx file
     WritePCXfile (lbmname, linear,
 		  SCREENWIDTH, SCREENHEIGHT,
 		  W_CacheLumpName ("PLAYPAL",PU_CACHE));
-	
+
     players[consoleplayer].message = "screen shot";
 }
